:sparkles: FEAT. 놀이터 후기 작성 기능 추가

+++ src/main/java/com/likelion/RePlay/domain/learning/entity/LearningReview.java
@@ -24,7 +24,7 @@ public class LearningReview extends BaseEntity {
     private String content;
 
     @Column(name="RATE")
-    private BigDecimal rate;
+    private Double rate;
 
     @ManyToOne
     @JoinColumn(name = "USER_ID")
